+++ src/composables/nodePack/useInstalledPacks.ts
@@ -1,6 +1,6 @@
 import { computed, onUnmounted } from 'vue'
 
-import { useNodePacks } from '@/composables/useNodePacks'
+import { useNodePacks } from '@/composables/nodePack/useNodePacks'
 import { useComfyManagerStore } from '@/stores/comfyManagerStore'
 import { UseNodePacksOptions } from '@/types/comfyManagerTypes'
 import type { components } from '@/types/comfyRegistryTypes'
